main.js: TODO B C and D
TODO B: When typing in the "title" field, we want to auto-complete based on article titles that already exist. You may use the api.php?prefixsearch={search} endpoint for auto-completion.

TODO C: When the user selects an item from the auto-complete list, we want
the textarea to populate with that article's contents.

TODO D: The autocomplete list should only be shown when the input receives
focus.