style.css: TODO A and TODO B
 TODO A: Modify or add styles to make the page look *roughly* like the design
spec file at `design-spec/layout.png`. Don't worry about making it pixel perfect
or matching the colors/spacing exactly. Instead, focus on the alignment of
elements. Elements that appear to be horizontally or vertically centered in the
spec, should be made that way here.

TODO B: Ensure that the page is responsive for tablet and mobile.